# backend/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Bitácora SOC API
  description: API completa para el sistema de Bitácora SOC con gestión de entradas, checklist de turno, usuarios y reportes
  version: 1.0.0
  contact:
    name: Bitácora SOC Team
servers:
  - url: http://192.168.100.50:3000/api
    description: Servidor de desarrollo (IP)
  - url: http://localhost:3000/api
    description: Servidor local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /config:
    get:
      tags:
        - Configuración
      summary: Obtener configuración
      description: Obtiene la configuración global de la aplicación
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Configuración actual
    
    put:
      tags:
        - Configuración
      summary: Actualizar configuración
      description: Actualiza la configuración global (solo admin)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                guestModeEnabled:
                  type: boolean
                guestMaxDurationDays:
                  type: integer
                  minimum: 1
                  maximum: 30
                shiftCheckCooldownHours:
                  type: integer
                  minimum: 1
                  maximum: 1440
                logoUrl:
                  type: string
      responses:
        '200':
          description: Configuración actualizada
...